+++ app/_definitions/models/entities/CbsContractFundingBudget.ts
@@ -50,7 +50,7 @@ const entity: RapidEntity<TEntitySingularCodes, TDictionaryCodes> = {
     {
       code: 'totalAmount',
       name: '合同金额',
-      type: 'integer',
+      type: 'double',
       required: true,
     },
     {
