+++ app/_definitions/models/entities/CbsOrderItem.ts
@@ -43,13 +43,13 @@ const entity: RapidEntity<TEntitySingularCodes, TDictionaryCodes> = {
     {
       code: 'price',
       name: '单价',
-      type: 'integer',
+      type: 'double',
       required: true,
     },
     {
       code: 'quantity',
       name: '数量',
-      type: 'integer',
+      type: 'double',
       required: true,
     },
     {
@@ -62,7 +62,7 @@ const entity: RapidEntity<TEntitySingularCodes, TDictionaryCodes> = {
     {
       code: 'taxRate',
       name: '税率',
-      type: 'integer',
+      type: 'double',
       required: true,
     },
   ],
